rename op paddle.fluid.matmul to paddle.matmul
modify original API fluid.layers.matmul,
and add new API paddle.matmul. test=develop
