+++ python/setup.py.in
@@ -111,6 +111,7 @@ packages=['paddle',
           'paddle.reader',
           'paddle.distributed',
           'paddle.fluid',
+          'paddle.tensor',
           'paddle.fluid.dygraph',
           'paddle.fluid.dygraph.dygraph_to_static',
           'paddle.fluid.proto',
